requirements.txt
icecream
pillow
prompt-toolkit # for lizard monitor
pynmea2
geopy
shapely
fiona
geopandas
rosys == v0.18.6
uvicorn == 0.28.1
mcap == 1.2.1
mcap-ros2-support == 0.5.5